Looks like the exception is an artifact of unittesting
I'm not entirely sure if it is due to Python or LLVM changes that reveal this.

However, the key is that these qjit kernels need to be at Python global scope.

For unit testing purposes, we put the qjit inside the test case body, just need to explicitly specify global.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>